requirements.txt
--extra-index-url https://backend:dc222fa6-452f-4253-a2fd-a52e7ba2e7e9@nexus.sigame.com.br/repository/pypi/simple
heimdall-client==2.2.0
aioboto3==9.6.0
Flask==2.1.2
motor==2.5.1
pydantic==1.9.0
python-decouple==3.6

uvloop==0.17.0
asgiref==3.5.2
hypercorn==0.14.3
uvicorn==0.20.0